+++ src/commands/diffWithRevision.ts
@@ -45,7 +45,7 @@ export class DiffWithRevisionCommand extends ActiveEditorCommand {
         try {
             const log = await Container.git.getLogForFile(gitUri.repoPath, gitUri.fsPath, {
                 maxCount: args.maxCount,
-                ref: (args.branchOrTag && args.branchOrTag.name) || gitUri.sha
+                ref: (args.branchOrTag && args.branchOrTag.ref) || gitUri.sha
             });
             if (log === undefined) {
                 if (args.branchOrTag) {